Add the dependencies for war
