+++ gia-rest/src/main/java/iisi/example/gia/batchconfig/BatchConstants.java
@@ -1,4 +1,4 @@
-package iisi.example.gia.config;
+package iisi.example.gia.batchconfig;
 
 public class BatchConstants {
