+++ gia-rest/src/main/java/iisi/example/gia/batchconfig/EmpXMLExportBatchConfig.java
@@ -1,4 +1,4 @@
-package iisi.example.gia.config;
+package iisi.example.gia.batchconfig;
 
 import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.context.annotation.Configuration;
