+++ gia-thymeleaf/pom.xml
@@ -11,6 +11,7 @@
 
     <groupId>iisi</groupId>
     <artifactId>gia-thymeleaf</artifactId>
+    <packaging>war</packaging>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
@@ -71,4 +72,14 @@
             <version>3.0.3</version>
         </dependency>
     </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-maven-plugin</artifactId>
+            </plugin>
+        </plugins>
+    </build>
+
 </project>
