+++ gia-thymeleaf/src/main/java/iisi/example/gia/ThymeleafApplication.java
@@ -1,0 +1,22 @@
+package iisi.example.gia;
+
+import org.mybatis.spring.annotation.MapperScan;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+@SpringBootApplication
+@MapperScan
+public class ThymeleafApplication extends SpringBootServletInitializer {
+
+	@Override
+	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+		return application.sources(ThymeleafApplication.class);
+	}
+
+	public static void main(String[] args) {
+		SpringApplication.run(ThymeleafApplication.class, args);
+	}
+
+}
