+++ gia-zk/pom.xml
@@ -10,6 +10,7 @@
     </parent>
 
     <artifactId>gia-zk</artifactId>
+    <packaging>war</packaging>
 
     <properties>
         <maven.compiler.source>17</maven.compiler.source>
@@ -37,6 +38,7 @@
     </repositories>
 
     <dependencies>
+
         <dependency>
             <groupId>iisi</groupId>
             <artifactId>gia-model</artifactId>
@@ -89,5 +91,14 @@
             <version>${zk.version}</version>
         </dependency>
     </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-maven-plugin</artifactId>
+            </plugin>
+        </plugins>
+    </build>
 
 </project>
