+++ gia-zk/src/main/java/iisi/example/gia/ZKApplication.java
@@ -1,0 +1,22 @@
+package iisi.example.gia;
+
+import org.mybatis.spring.annotation.MapperScan;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+@SpringBootApplication
+@MapperScan
+public class ZKApplication extends SpringBootServletInitializer {
+
+    @Override
+    protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+        return application.sources(ZKApplication.class);
+    }
+
+    public static void main(String[] args) {
+        SpringApplication.run(ZKApplication.class, args);
+    }
+
+}
