+++ gia-zk/src/main/java/iisi/example/gia/vm/EmpManageVM.java
@@ -10,7 +10,6 @@ import org.zkoss.zk.ui.select.annotation.VariableResolver;
 import org.zkoss.zk.ui.select.annotation.WireVariable;
 import org.zkoss.zkplus.spring.DelegatingVariableResolver;
 import org.zkoss.zul.Messagebox;
-
 import java.util.List;
 import java.util.Set;
