+++ gia-zk/src/main/resources/application.properties
@@ -20,4 +20,4 @@ mybatis.mapper-locations=classpath:mappers/*.xml
 mybatis.configuration.log-impl=org.apache.ibatis.logging.stdout.StdOutImpl
 
 spring.thymeleaf.cache=false
-zk.homepage=hello
+zk.homepage=empManage
